requirements.txt
kafka-python
pyspark
pytest
pyyaml
duckdb
streamlit
pandas
